+++ sorc/derived/atmos/atmos.py
@@ -1,20 +1,9 @@
 # =========================================================================
-
-# Module: derived/atmos/atmos.py
-
+# File: sorc/derived/atmos/atmos.py
 # Author: Henry R. Winterbottom
-
-# Email: henry.winterbottom@noaa.gov
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the respective public license published by the
-# Free Software Foundation and included with the repository within
-# which this application is contained.
-
-# This program is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-
+# Date: 29 August 2023
+# Version: 0.0.1
+# License: LGPL v2.1
 # =========================================================================
 
 """
@@ -57,8 +46,7 @@ __email__ = "henry.winterbottom@noaa.gov"
 
 # ----
 
-from derived import Derived
-
+from derived.derived import Derived
 
 from types import SimpleNamespace
