+++ sorc/derived/atmos/pressures.py
@@ -17,6 +17,14 @@
 
 # =========================================================================
 
+# =========================================================================
+# File: sorc/derived/atmos/pressures.py
+# Author: Henry R. Winterbottom
+# Date: 29 August 2023
+# Version: 0.0.1
+# License: LGPL v2.1
+# =========================================================================
+
 """
 Module
 ------
